+++ app/models/parser.py
@@ -1,0 +1,31 @@
+from enum import Enum
+
+from pydantic import BaseModel, TypeAdapter
+
+
+class ParserFunction(str, Enum):
+    DSPIPELINES = "dspipelines"
+    VESPUCCI = "vespucci"
+
+
+class ParserSubgraphLine(BaseModel):
+    start: int
+    end: int
+
+
+class ParserSubgraphCursor(BaseModel):
+    start: int
+    end: int
+
+
+class ParserSubgraph(BaseModel):
+    id: str
+    library: str
+    function: str
+    source: str
+    step_name: str
+    line: ParserSubgraphLine
+    cursor: ParserSubgraphCursor
+
+
+ParserSubgraphList = TypeAdapter(list[ParserSubgraph])