+++ app/profiling_functions/dspipelines.py
@@ -1,9 +1,9 @@
 import json
-
 from pathlib import Path
 
-from app.custom_types import ParserSubgraph, SupportedTaxonomiesFunction
-from app.profiling_functions._base import BaseMLProfiler
+from app.models.parser import ParserSubgraph
+from app.models.profiler import ProfileResult
+from app.profiling_functions.base import BaseMLProfiler, Taxonomy
 
 steps_functions_mapping_path = Path(
     "./resources/taxonomies/extra/dspipelines_steps_functions_mapping.json"
@@ -14,17 +14,9 @@ steps_functions_mapping: dict[str, str] = json.loads(
 
 
 class DSPipelinesProfiler(BaseMLProfiler):
+    def __init__(self, source_code: str, taxonomy: Taxonomy):
+        super().__init__(source_code, taxonomy)
 
-    def __init__(self, python_content: str, taxonomy_name: SupportedTaxonomiesFunction):
-        super().__init__(python_content, taxonomy_name)
-
-    def profile_subgraph(
-        self, subgraph: ParserSubgraph, default_step: str
-    ) -> tuple[str, float | None, list[list[tuple[str, float]]]]:
-        retrieved_step = steps_functions_mapping.get(subgraph.function, default_step)
-        verified_retrieved_step = (
-            retrieved_step
-            if retrieved_step in self.taxonomy.get_original_steps_names()
-            else default_step
-        )
-        return (verified_retrieved_step, 1, [[(verified_retrieved_step, 1.0)]])
+    async def profile_subgraph(self, subgraph: ParserSubgraph) -> ProfileResult:
+        retrieved_step = steps_functions_mapping.get(subgraph.function, self.taxonomy.default_step)
+        return ProfileResult(step=retrieved_step, perplexity=1)